Moved styles out of being local in About
