+++ src/components/Profile/ProfileName.jsx
@@ -2,7 +2,7 @@
 import IconButton from '@material-ui/core/IconButton';
 import AddAPhotoIcon from '@material-ui/icons/AddAPhoto';
 import CheckCircleIcon from '@material-ui/icons/CheckCircle';
-import useStyles from './EditVendorProfile.styles';
+import useStyles from './Profile.styles';
 import { Grid, Box, Typography, Avatar } from '@material-ui/core';
 
 import starsImg from '../Images/stars.jpg';
